+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## HEAD
 
+* Give no indication of unknown account on request of password reset email.
 * Use 422 status code for form submission error responses.
 * Make default cookie name depend on Rails environment.